Backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.3
beautifulsoup4>=4.12
requests>=2.31
lxml>=4.9
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
statsmodels>=0.14
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
gunicorn>=21.2
